common functions exported to UserUtils class
