+++ src/socket/events/load_current_course.ts
@@ -2,12 +2,7 @@ import { JwtPayload } from "jsonwebtoken";
 import { secretToken } from "../../const/config";
 import Boom from "boom";
 import { io } from "socket.io-client";
-import {
-    getProgressForVocabulary,
-  hasUserAchievedWithoutRequest,
-  hasUserCompletedLessonToday,
-  hasUserCompletedLessonYesterday,
-} from "../../const/common";
+
 import { ErrorCodes } from "../../const/error_codes";
 import { UserCourseEntity } from "../../entities/user_course_entity";
 import { UserEntity } from "../../entities/user_entity";
@@ -17,6 +12,7 @@ import { TopicService } from "../../services/topic_service";
 import { UserCourseService } from "../../services/user_course_service";
 import { UserService } from "../../services/user_service";
 import { UserStatisticsService } from "../../services/user_statistics_service";
+import { UserUtils } from "../../logic/user";
 
 const jwt = require("jsonwebtoken");
 export async function loadCurrentCourse(
@@ -38,7 +34,7 @@ export async function loadCurrentCourse(
 
   if (activeCoursesData.length == 0) {
   }
-  var activeCourses = await getProgressForVocabulary(activeCoursesData, user);
+  var activeCourses = await UserUtils.getProgressForVocabulary(activeCoursesData, user);
   var activeCourse;
   activeCourses!.forEach((e) => {
     if (e.activeCourse.userCourse.id == user.profile.userCourse.id) {
